refactor : typegard/index.ts에서 typegard디렉토리에 포함된 파일을 내보내도록 수정 1. index.ts파일 추가 2. typegard에 정의되었던 파일을 import해왔던 파일의 경로 수정
